lagt flyttet useInterval utenfor modalWrapper, og oppdatert default state i dev
