+++ src/server/server.ts
@@ -67,7 +67,6 @@ app.use((req, res, next) => {
 // Metrics
 app.use(
     createMiddleware({
-        // @ts-ignore
         app,
         options: {
             labels: ['app', 'namespace', 'cluster'],
